# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  internal_http_handlers.APIError:
    properties:
      code:
        type: string
      details: {}
      message:
        type: string
    type: object
...
  internal_http_handlers.LoginResponse:
    properties:
      access_token:
        type: string
      message:
        type: string
      refresh_token:
        type: string
      user:
        properties:
          email:
            type: string
          id:
            type: integer
          role:
            type: string
        type: object
    type: object
  internal_http_handlers.LogoutRequest:
    properties:
      refresh_token:
        type: string
    type: object
  internal_http_handlers.LogoutResponse:
    properties:
      message:
        type: string
    type: object
  internal_http_handlers.RefreshTokenRequest:
    properties:
      refresh_token:
        type: string
    required:
    - refresh_token
    type: object
...
  internal_http_handlers.loginRequest:
    properties:
      email:
        maxLength: 255
        type: string
      password:
        maxLength: 128
        minLength: 8
        type: string
    required:
    - email
    - password
    type: object
  internal_http_handlers.registerRequest:
    properties:
      email:
        maxLength: 255
        type: string
      password:
        maxLength: 128
        minLength: 8
        type: string
    required:
    - email
    - password
    type: object
host: localhost:8080
info:
  contact:
    email: support@diana-health.com
    name: DIANA Support
  description: Diabetes Insulin Activity and Nutrition Analyzer (DIANA) - Risk assessment
    and patient management API
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: DIANA API
  version: "1.0"
paths:
...
  /auth/login:
    post:
      consumes:
      - application/json
      description: Authenticate user and return JWT tokens
      parameters:
      - description: Login credentials
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_http_handlers.loginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Login successful
          schema:
            $ref: '#/definitions/internal_http_handlers.LoginResponse'
        "400":
          description: Invalid request payload
          schema:
            $ref: '#/definitions/internal_http_handlers.APIError'
        "401":
          description: Invalid credentials
          schema:
            $ref: '#/definitions/internal_http_handlers.APIError'
        "403":
          description: Account inactive
          schema:
            $ref: '#/definitions/internal_http_handlers.APIError'
      summary: Login user
      tags:
      - Authentication
  /auth/logout:
    post:
      consumes:
      - application/json
      description: Revoke refresh token and clear authentication cookies
      parameters:
      - description: Optional refresh token to revoke
        in: body
        name: request
        schema:
          $ref: '#/definitions/internal_http_handlers.LogoutRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Logged out successfully
          schema:
            $ref: '#/definitions/internal_http_handlers.LogoutResponse'
      summary: Logout user
      tags:
      - Authentication
  /auth/refresh:
    post:
      consumes:
      - application/json
      description: Generate new access token using refresh token
      parameters:
      - description: Refresh token
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_http_handlers.RefreshTokenRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Token refreshed successfully
          schema:
            $ref: '#/definitions/internal_http_handlers.LoginResponse'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/internal_http_handlers.APIError'
        "401":
          description: Invalid or expired refresh token
          schema:
            $ref: '#/definitions/internal_http_handlers.APIError'
      summary: Refresh access token
      tags:
      - Authentication
  /auth/register:
    post:
      consumes:
      - application/json
      description: Create a new user account
      parameters:
      - description: Registration details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_http_handlers.registerRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Registration successful
          schema:
            $ref: '#/definitions/internal_http_handlers.LoginResponse'
        "400":
          description: Invalid input
          schema:
            $ref: '#/definitions/internal_http_handlers.APIError'
        "409":
          description: Email already exists
          schema:
            $ref: '#/definitions/internal_http_handlers.APIError'
      summary: Register new user
      tags:
      - Authentication
...
securityDefinitions:
  BearerAuth:
    description: 'JWT Bearer token. Format: Bearer <token>'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"